+++ src/Domain/Api.Ai.Domain.DataTransferObject/Response/Message/TextMessageResponse.cs
@@ -33,7 +33,7 @@ namespace Api.Ai.Domain.DataTransferObject.Response.Message
 
         public override void SetMessageType()
         {
-            this.Type = Enum.Type.Text;
+            this.Type = (int)Enum.Type.Text;
         }
 
         #endregion
